add correct label to ds config initContainer
CRI-O 1.16 made a change to how selinux labels were handled that increased security. In doing so, pod specs need to enable selinuxRelabel. In the case of this initContainer, it is trying to edit the permissions of a file on the host system, which is a security flaw, and thus selinux is denying it.

Add the type spc_t, which allows the container to modify the host file

Signed-off-by: Peter Hunt <pehunt@redhat.com>