auditConfig: Log write operations of Identities with RequestResponse
In order to be able to match a user from IDP with a username which is
normally logged in the audit logs, we need to have the creation of the
Identity API resource logged.

This patch changes the hardcoded auditConfig to log any write
operation (create, update, patch) of the "identities" resource of the
"user.openshift.io" group at the RequestResponse level.