Add default ClusterRole for motd
With https://github.com/openshift/oc/pull/67 came the addition of the
capability to display a message of the day via the `oc` command. This
message of the day serves the purpose of presenting a legal notice
that's required for regulatory compliance.

In order for this legal notice to be usable out of the box, we need to
grant folks read-only access to it. This PR adds that by creating a
default ClusterRole and ClusterRoleBinding that enables this.