splits the encryption tests
The current timeout for all encryption test is 4 hours.
Previously the time needed to roll out a new KAS was 15 minutes, due to a known issue on AWS with the NLB it can take now 22 minutes.
It means that the tests need ~25 min to make sure that all servers are on the same revision and it is safe to start the test.

Since the rotation scenario needs to enable the encryption first, then wait and check if all resources were migrated, then change the encryption key and check the migration again.
It means that only these tests will need ~1 hour just for installing the servers.

Thus this PR moves the rotation tests to a separate CI job.
It will shorten the overall time needed to test encryption but also increase the pass rate of the tests as we have seen them frequently failing to due to a timeout.