fix: TestRedeployOnConfigChange
Previously, the test modified the etcd-signer-ca to see if a new re-deployment would be triggered.
About two months ago, the CEO introduced a change that runs a controller which overwrites (syncs) the secret from a different source.

This change broke the test, which has been failing for nearly two months.

This PR changes the test to modify the operator's configuration,
which is fully controlled by the operator, to trigger a re-deployment.
This is what the test was originally supposed to check.