MULTIARCH-4557: Sync import mode image config status field in the observed config
As per https://github.com/openshift/enhancements/pull/1605, there was a
new field introduced in the image config spec and status which reflects
the global value to be set for imagestream import mode which is behind a
featuregate. This PR reads the image config status, checks if the importmode
string is present and syncs that value in the imagepolicyconfig of the
observed config to be used by the apiserver.